Relocate chunker.rs under merge_batcher/
