+++ differential-dataflow/src/trace/implementations/mod.rs
@@ -42,10 +42,9 @@ pub mod spine_fueled;
 
 pub mod merge_batcher;
 pub mod ord_neu;
-pub mod chunker;
 
 // Opinionated takes on default spines.
-pub use self::chunker::ContainerChunker;
+pub use self::merge_batcher::chunker::ContainerChunker;
 pub use self::ord_neu::OrdValSpine as ValSpine;
 pub use self::ord_neu::OrdValBatcher as ValBatcher;
 pub use self::ord_neu::VecOrdValBuilder as ValBuilder;
